Revert "Investigation: asset image generation moved to franklin lib"
This reverts commit 40736730ebaad6ee775f406759fbd313181fa4ac.
