+++ scripts/lib-franklin.js
@@ -10,8 +10,6 @@
  * governing permissions and limitations under the License.
  */
 
-import { findNodesByRegExp } from './utils.js';
-
 /**
  * log RUM if part of the sample.
  * @param {string} checkpoint identifies the checkpoint in funnel
@@ -443,7 +441,6 @@ export async function loadBlocks(main) {
 export function createOptimizedPicture(src, alt = '', eager = false, breakpoints = [{ media: '(min-width: 400px)', width: '2000' }, { width: '750' }]) {
   const url = new URL(src, window.location.href);
   const picture = document.createElement('picture');
-  picture.classList.add('groundfog');
   const { pathname } = url;
   const ext = pathname.substring(pathname.lastIndexOf('.') + 1);
 
@@ -611,88 +608,6 @@ export function setup() {
     }
   }
 }
-
-export const imagePathTemplate = ({
-  uuid, width, quality = 'high', extension = 'webp',
-}) => `https://assets.new.siemens.com/siemens/assets/api/uuid:${uuid}/width:${width}/quality:${quality}/${uuid}.${extension}`;
-
-export const renditionsWidths = [
-  3840,
-  2732,
-  2224,
-  2048,
-  1920,
-  1266,
-  1125,
-  750,
-  640,
-  320,
-  100,
-];
-
-export const createAssetPicture = (uuid, alt = '', eager = false) => {
-  const picture = document.createElement('picture');
-  const source = document.createElement('source');
-  const renditions = {
-    webp: [],
-    jpg: [],
-  };
-
-  for (let i = 0; i < renditionsWidths.length - 1; i += 1) {
-    const width = renditionsWidths[i];
-    renditions.jpg.push(`${imagePathTemplate({
-      uuid,
-      width,
-      extension: 'jpg',
-    })} ${width}w`);
-    renditions.webp.push(`${imagePathTemplate({
-      uuid,
-      width,
-    })} ${width}w`);
-  }
-  source.setAttribute('srcset', [
-    ...renditions.webp,
-    ...renditions.jpg,
-  ].join(','));
-  picture.appendChild(source);
-
-  const img = document.createElement('img');
-  img.setAttribute('loading', eager ? 'eager' : 'lazy');
-  img.setAttribute('alt', alt);
-  picture.appendChild(img);
-  img.setAttribute('src', imagePathTemplate({
-    uuid,
-    width: renditionsWidths[renditionsWidths.length - 1],
-    extension: 'jpg',
-    quality: 'low',
-  }));
-
-  return picture;
-};
-
-/**
- * Decorates media from Siemens Assets.
- * @param {Element} element Container element
- */
-export const decorateMedia = (element) => {
-  const MEDIA_REGEX = /image|video:([a-z0-9-]*)/gi;
-  findNodesByRegExp(element, MEDIA_REGEX).forEach((media) => {
-    const mediaMarker = media.textContent.split(':');
-
-    switch (mediaMarker[0].toLowerCase()) {
-      case 'image':
-        media.textContent = '';
-        media.append(createAssetPicture(mediaMarker[1]));
-        break;
-      case 'video':
-        break;
-
-      default:
-        // eslint-disable-next-line
-        console.log(`Unsupported media markup: ${media.textContent}`);
-    }
-  });
-};
 
 /**
  * Auto initializiation.
