+++ scripts/scripts.js
@@ -11,10 +11,9 @@ import {
   waitForLCP,
   loadBlocks,
   loadCSS,
-  decorateMedia,
 } from './lib-franklin.js';
 
-import { decorateAsFluidwebPage } from './fluidweb.js';
+import { decorateAsFluidwebPage, decorateMedia } from './fluidweb.js';
 
 const LCP_BLOCKS = []; // add your LCP blocks to the list
 window.hlx.RUM_GENERATION = 'project-1'; // add your RUM generation information here
